+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Add balenaEtcher to the minimal image.
 - Add 'asusctl-git' as a pre-built package in the winesapOS repository.
 - Add NonSteamLaunchers to manage other game launchers.
+- Add Nexus Mods app for managing mods.
 
 ## 4.1.0 - The Dual-Boot Update